+++ classes/phing/tasks/ext/phploc/PHPLocTask.php
@@ -20,6 +20,8 @@
 
 require_once 'phing/Task.php';
 require_once 'phing/BuildException.php';
+require_once 'phing/tasks/ext/phploc/PHPLocFormatterElement.php';
+require_once 'phing/tasks/ext/phploc/PHPLocFormatterFactory.php';
 
 /**
  * Runs phploc a tool for quickly measuring the size of PHP projects.
@@ -75,16 +77,9 @@ class PHPLocTask extends Task
     protected $fileSets = array();
 
     /**
-     * @var bool
+     * @var PHPLocFormatterElement[]
      */
-    protected $oldVersion = false;
-
-    /**
-     * Flag for identifying if phploc library is between version 1.7.0 and 1.7.4
-     *
-     * @var bool
-     */
-    private $isOneSevenVersion = false;
+    protected $formatterElements = array();
 
     /**
      * @var string
@@ -164,6 +159,14 @@ class PHPLocTask extends Task
     }
 
     /**
+     * @param PHPLocFormatterElement $formatterElement
+     */
+    public function addFormatter(PHPLocFormatterElement $formatterElement)
+    {
+        $this->formatterElements[] = $formatterElement;
+    }
+
+    /**
      * @throws BuildException
      */
     protected function loadDependencies()
@@ -183,24 +186,11 @@ class Application
 
         if (!class_exists('\SebastianBergmann\PHPLOC\Analyser')) {
             if (!@include_once 'SebastianBergmann/PHPLOC/autoload.php') {
-                if (!@include_once 'PHPLOC/Analyser.php') {
-                    throw new BuildException(
-                        'PHPLocTask depends on PHPLoc being installed and on include_path.',
-                        $this->getLocation()
-                    );
-                } else {
-                    $this->oldVersion = true;
-                }
+                throw new BuildException(
+                    'PHPLocTask depends on PHPLoc being installed and on include_path.',
+                    $this->getLocation()
+                );
             }
-        }
-
-        $versionClass = '\\SebastianBergmann\\PHPLOC\\Version';
-
-        if (class_exists($versionClass)
-            && version_compare(constant($versionClass . '::VERSION'), '1.7.0') >= 0
-            && version_compare(constant($versionClass . '::VERSION'), '2.0.0beta1') == -1
-        ) {
-            $this->isOneSevenVersion = true;
         }
     }
 
@@ -209,20 +199,6 @@ class Application
         $this->loadDependencies();
 
         $this->validateProperties();
-
-        if ($this->reportDirectory !== null && !is_dir($this->reportDirectory)) {
-            $reportOutputDir = new PhingFile($this->reportDirectory);
-
-            $logMessage = "Report output directory doesn't exist, creating: "
-                . $reportOutputDir->getAbsolutePath() . '.';
-
-            $this->log($logMessage);
-            $reportOutputDir->mkdirs();
-        }
-
-        if ($this->reportType !== 'cli') {
-            $this->reportFileName .= '.' . $this->reportType;
-        }
 
         if (count($this->fileSets) > 0) {
             foreach ($this->fileSets as $fileSet) {
@@ -270,16 +246,39 @@ class Application
             throw new BuildException('No file suffix defined.');
         }
 
-        if ($this->reportType === null) {
-            throw new BuildException('No report type defined.');
-        }
+        if (count($this->formatterElements) == 0) {
+            if ($this->reportType === null) {
+                throw new BuildException('No report type or formatters defined.');
+            }
 
-        if ($this->reportType !== null && !in_array($this->reportType, $this->acceptedReportTypes)) {
-            throw new BuildException('Unaccepted report type defined.');
-        }
+            if ($this->reportType !== null && !in_array($this->reportType, $this->acceptedReportTypes)) {
+                throw new BuildException('Unaccepted report type defined.');
+            }
 
-        if ($this->reportType !== 'cli' && $this->reportDirectory === null) {
-            throw new BuildException('No report output directory defined.');
+            if ($this->reportType !== 'cli' && $this->reportDirectory === null) {
+                throw new BuildException('No report output directory defined.');
+            }
+
+            if ($this->reportDirectory !== null && !is_dir($this->reportDirectory)) {
+                $reportOutputDir = new PhingFile($this->reportDirectory);
+
+                $logMessage = "Report output directory doesn't exist, creating: "
+                    . $reportOutputDir->getAbsolutePath() . '.';
+
+                $this->log($logMessage);
+                $reportOutputDir->mkdirs();
+            }
+
+            if ($this->reportType !== 'cli') {
+                $this->reportFileName .= '.' . $this->reportType;
+            }
+
+            $formatterElement = new PHPLocFormatterElement();
+            $formatterElement->setType($this->reportType);
+            $formatterElement->setUseFile($this->reportDirectory !== null);
+            $formatterElement->setToDir($this->reportDirectory);
+            $formatterElement->setOutfile($this->reportFileName);
+            $this->formatterElements[] = $formatterElement;
         }
     }
 
@@ -298,94 +297,17 @@ class Application
         $files = $this->getFilesToCheck();
         $count = $this->getCountForFiles($files);
 
-        if ($this->reportType != 'cli') {
-            $logMessage = 'Writing report to: '
-                . $this->reportDirectory . DIRECTORY_SEPARATOR . $this->reportFileName;
+        foreach ($this->formatterElements as $formatterElement) {
+            $formatter = PHPLocFormatterFactory::createFormatter($formatterElement);
 
-            $this->log($logMessage);
-        }
+            if ($formatterElement->getType() != 'cli') {
+                $logMessage = 'Writing report to: '
+                    . $formatterElement->getToDir() . DIRECTORY_SEPARATOR . $formatterElement->getOutfile();
 
-        switch ($this->reportType) {
-            case 'cli':
-                if ($this->oldVersion || $this->isOneSevenVersion) {
-                    if ($this->oldVersion) {
-                        require_once 'PHPLOC/TextUI/ResultPrinter/Text.php';
+                $this->log($logMessage);
+            }
 
-                        $printerClass = 'PHPLOC_TextUI_ResultPrinter_Text';
-                    } else {
-                        $printerClass = '\\SebastianBergmann\\PHPLOC\\TextUI\\ResultPrinter';
-                    }
-
-                    $printer = new $printerClass();
-                    $printer->printResult($count, $this->countTests);
-                } else {
-                    $outputClass = '\\Symfony\\Component\\Console\\Output\\ConsoleOutput';
-                    $printerClass = '\\SebastianBergmann\\PHPLOC\\Log\\Text';
-
-                    $output = new $outputClass();
-                    $printer = new $printerClass();
-                    $printer->printResult($output, $count, $this->countTests);
-                }
-                break;
-
-            case 'txt':
-                if ($this->oldVersion || $this->isOneSevenVersion) {
-                    if ($this->oldVersion) {
-                        require_once 'PHPLOC/TextUI/ResultPrinter/Text.php';
-
-                        $printerClass = 'PHPLOC_TextUI_ResultPrinter_Text';
-                    } else {
-                        $printerClass = '\\SebastianBergmann\\PHPLOC\\TextUI\\ResultPrinter';
-                    }
-
-                    $printer = new $printerClass();
-
-                    ob_start();
-                    $printer->printResult($count, $this->countTests);
-                    $result = ob_get_contents();
-                    ob_end_clean();
-
-                    file_put_contents($this->reportDirectory . DIRECTORY_SEPARATOR . $this->reportFileName, $result);
-                } else {
-                    $outputClass = '\\Symfony\\Component\\Console\\Output\\StreamOutput';
-                    $printerClass = '\\SebastianBergmann\\PHPLOC\\Log\\Text';
-
-                    $stream = fopen($this->reportDirectory . DIRECTORY_SEPARATOR . $this->reportFileName, 'a+');
-                    $output = new $outputClass($stream);
-                    $printer = new $printerClass();
-                    $printer->printResult($output, $count, $this->countTests);
-                }
-                break;
-
-            case 'xml':
-                if ($this->oldVersion) {
-                    require_once 'PHPLOC/TextUI/ResultPrinter/XML.php';
-
-                    $printerClass = 'PHPLOC_TextUI_ResultPrinter_XML';
-                } else {
-                    $printerClass = '\\SebastianBergmann\\PHPLOC\\Log\\XML';
-                }
-
-                $printer = new $printerClass();
-                $printer->printResult($this->reportDirectory . DIRECTORY_SEPARATOR . $this->reportFileName, $count);
-                break;
-
-            case 'csv':
-                if ($this->oldVersion) {
-                    require_once 'PHPLOC/TextUI/ResultPrinter/CSV.php';
-
-                    $printerClass = 'PHPLOC_TextUI_ResultPrinter_CSV';
-                } else {
-                    if ($this->isOneSevenVersion) {
-                        $printerClass = '\\SebastianBergmann\\PHPLOC\\Log\\CSV';
-                    } else {
-                        $printerClass = '\\SebastianBergmann\\PHPLOC\\Log\\CSV\\Single';
-                    }
-                }
-
-                $printer = new $printerClass();
-                $printer->printResult($this->reportDirectory . DIRECTORY_SEPARATOR . $this->reportFileName, $count);
-                break;
+            $formatter->printResult($count, $this->countTests);
         }
     }
 
@@ -414,7 +336,7 @@ class Application
      */
     protected function getCountForFiles(array $files)
     {
-        $analyserClass = ($this->oldVersion ? 'PHPLOC_Analyser' : '\\SebastianBergmann\\PHPLOC\\Analyser');
+        $analyserClass = '\\SebastianBergmann\\PHPLOC\\Analyser';
         $analyser = new $analyserClass();
 
         return $analyser->countFiles($files, $this->countTests);