+++ classes/phing/tasks/ext/phpunit/PHPUnitTask.php
@@ -88,7 +88,9 @@ class PHPUnitTask extends Task
          * PEAR old-style, then composer, then PHAR
          */
         @include_once 'PHPUnit/Runner/Version.php';
-        @include_once 'phpunit/Runner/Version.php';
+        if (!class_exists('PHPUnit_Runner_Version')) {
+            @include_once 'phpunit/Runner/Version.php';
+        }
         if (!empty($this->pharLocation)) {
             $GLOBALS['_SERVER']['SCRIPT_NAME'] = '-';
             ob_start();
@@ -390,12 +392,14 @@ class PHPUnitTask extends Task
             }
         }
 
-        $browsers = $config->getSeleniumBrowserConfiguration();
+        if (method_exists($config, 'getSeleniumBrowserConfiguration')) {
+            $browsers = $config->getSeleniumBrowserConfiguration();
 
-        if (!empty($browsers) &&
-            class_exists('PHPUnit_Extensions_SeleniumTestCase')
-        ) {
-            PHPUnit_Extensions_SeleniumTestCase::$browsers = $browsers;
+            if (!empty($browsers) &&
+                class_exists('PHPUnit_Extensions_SeleniumTestCase')
+            ) {
+                PHPUnit_Extensions_SeleniumTestCase::$browsers = $browsers;
+            }
         }
 
         return $phpunit;
@@ -497,7 +501,9 @@ class PHPUnitTask extends Task
             $path = realpath($pwd . '/../../../');
 
             $filter = new PHP_CodeCoverage_Filter();
-            $filter->addDirectoryToBlacklist($path);
+            if (method_exists($filter, 'addDirectoryToBlacklist')) {
+                $filter->addDirectoryToBlacklist($path);
+            }
             $runner->setCodecoverage(new PHP_CodeCoverage(null, $filter));
         }