+++ test/classes/phing/regression/PearPkg2CompatibilityTest.php
@@ -36,14 +36,19 @@ class PearPkg2CompatibilityTest extends BuildFileTest
 
     public function setUp()
     {
-        if (defined('HHVM_VERSION')) {
-            $this->markTestSkipped("PEAR tests do not run on HHVM");
-        }
-
         $this->savedErrorLevel = error_reporting();
         error_reporting(E_ERROR);
         $buildFile = PHING_TEST_BASE . "/etc/regression/524/build.xml";
         $this->configureProject($buildFile);
+
+        if (defined('HHVM_VERSION')) {
+            $this->markTestSkipped("PEAR tests do not run on HHVM");
+        }
+
+        if (!class_exists('PEAR_PackageFileManager', false)) {
+            $this->markTestSkipped("This test requires PEAR_PackageFileManager to be installed");
+        }
+
         $this->executeTarget("setup");
     }